+++ packages/sdk/src/internal/dapp/data-service-dapp-messages.ts
@@ -50,6 +50,8 @@ export class DataServiceDappMessages implements DappMessages {
         addressTypes: command?.addressTypes?.map((addr) =>
           toAddressTypeDto(addr),
         ),
+        // tags: command.tags,
+        actions: command.actions,
       }),
     );
   }